src/requirements.txt
asgiref==3.3.1
Django==3.0.5
django-appconf==1.0.4
django-background-tasks==1.2.5
django-compat==1.0.15
django-video-encoding==1.0.0
djangorestframework==3.11.0
ffmpeg==1.4
gunicorn==20.0.4
Pillow==6.2.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
